split chat routing into intelligence-loop and helper partials
